+++ fixtures/futures/tests/bindings/test_futures.py
@@ -3,6 +3,7 @@ import unittest
 from datetime import datetime
 import asyncio
 import typing
+import futures
 
 def now():
     return datetime.now()
@@ -104,6 +105,61 @@ class TestFutures(unittest.TestCase):
             self.assertGreater(t_delta, 0.2)
 
         asyncio.run(test())
+
+    def test_foreign_async_trait_interface_methods(self):
+        class PyAsyncParser:
+            def __init__(self):
+                self.completed_delays = 0
+
+            async def as_string(self, delay_ms, value):
+                await asyncio.sleep(delay_ms / 1000.0)
+                return str(value)
+
+            async def try_from_string(self, delay_ms, value):
+                await asyncio.sleep(delay_ms / 1000.0)
+                if value == "force-unexpected-exception":
+                    raise RuntimeError("UnexpectedException")
+                try:
+                    return int(value)
+                except:
+                    raise ParserError.NotAnInt()
+
+            async def delay(self, delay_ms):
+                await asyncio.sleep(delay_ms / 1000.0)
+                self.completed_delays += 1
+
+            async def try_delay(self, delay_ms):
+                try:
+                    delay_ms = int(delay_ms)
+                except:
+                    raise ParserError.NotAnInt()
+                await asyncio.sleep(delay_ms / 1000.0)
+                self.completed_delays += 1
+
+        async def test():
+            trait_obj = PyAsyncParser()
+            self.assertEqual(await as_string_using_trait(trait_obj, 1, 42), "42")
+            self.assertEqual(await try_from_string_using_trait(trait_obj, 1, "42"), 42)
+            with self.assertRaises(ParserError.NotAnInt):
+                await try_from_string_using_trait(trait_obj, 1, "fourty-two")
+            with self.assertRaises(ParserError.UnexpectedError):
+                await try_from_string_using_trait(trait_obj, 1, "force-unexpected-exception")
+            await delay_using_trait(trait_obj, 1)
+            await try_delay_using_trait(trait_obj, "1")
+            with self.assertRaises(ParserError.NotAnInt):
+                await try_delay_using_trait(trait_obj, "one")
+
+            completed_delays_before = trait_obj.completed_delays
+            await cancel_delay_using_trait(trait_obj, 10)
+            # sleep long enough so that the `delay()` call would finish if it wasn't cancelled.
+            await asyncio.sleep(0.1)
+            # If the task was cancelled, then completed_delays won't have increased
+            self.assertEqual(trait_obj.completed_delays, completed_delays_before)
+
+
+        asyncio.run(test())
+        # check that all foreign future handles were released
+        self.assertEqual(len(futures.UNIFFI_FOREIGN_FUTURE_HANDLE_MAP), 0)
 
     def test_async_object_param(self):
         async def test():