+++ uniffi_bindgen/src/bindings/swift/templates/HandleMap.swift
@@ -30,5 +30,11 @@ fileprivate class UniffiHandleMap<T> {
             return obj
         }
     }
+
+    var count: Int {
+        get {
+            map.count
+        }
+    }
 }